fixed bug where log was not being sent by primary
The log sending code waits for a signal to know there's more entries to be sent.
We limit the amout of log entries sent at a single time to prevent starvation of
other tasks. The bug happened when the limit was reached and no more entries
were appended, thus the signal doesn't trigger and there are still entries to be
sent.
The code now checks for more entries before it starts waiting on the signal.